% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{AnyNA}
\alias{AnyNA}
\title{'Not Available' / Missing Values}
\usage{
AnyNA(x)
}
\description{
A simple version of the newer \code{anyNA} function which essentially
implements any(is.na(x)).
}
\keyword{internal}

