% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_money_market_rotation.R
\name{attribute_dividends_to_rotations}
\alias{attribute_dividends_to_rotations}
\title{Attribute dividends to rotation periods}
\usage{
attribute_dividends_to_rotations(dividends, rotations)
}
\arguments{
\item{dividends}{Data frame of dividend payments}

\item{rotations}{Data frame of rotation history}
}
\value{
Data frame with dividends attributed to correct rotation period
}
\description{
Dividends are attributed based on which ticker was held, not payment date.
A dividend posting after position close is attributed to the rotation that held it.
}
