% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_activity_linking.R
\name{batch_ignore_activities}
\alias{batch_ignore_activities}
\title{Batch ignore activities}
\usage{
batch_ignore_activities(activity_ids)
}
\arguments{
\item{activity_ids}{Character vector of activity IDs}
}
\value{
Logical TRUE if successful
}
\description{
Efficiently sets ignore_for_grouping flag for multiple activities.
}
