% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{beetle}
\alias{beetle}
\title{Dobson's Beetle Data}
\format{A data frame with 8 rows and 3 variables}
\source{
A. Dobson, \emph{An Introduction to Generalized Linear Models}, Chapman &
Hall/CRC, 2002.
}
\description{
The data give the number of flour beetles killed after five hour exposure
to the insecticide carbon disulphide at eight different concentrations.
}
\details{
\itemize{
  \item \code{ldose} Log dose of carbon disulphide.
  \item \code{y} Number of beetles subjected to insecticide.
  \item \code{n} Number of beetles killed.
}
}
\keyword{datasets}
