% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bladder}
\alias{bladder}
\title{Bladder volume data}
\format{A data frame with 184 rows and 3 columns.}
\source{
Brown, P. (1993) \emph{Measurement, Regression, and Calibration}. 
Oxford University Press.
}
\description{
A series of 23 women patients attending a urodynamic clinic were recruited 
for a study. After successful voiding of the bladder, sterile water was 
introduced in additions of 10, 15, and then 25 ml increments up to a final 
cumulative total of 175 ml. At each volume a measure of height (H) in mm and 
depth (D) in mm of largest ultrasound bladder images were taken. The product 
H × D was taken as a measure of liquid volume.
}
\details{
\itemize{
  \item \code{subject} The subject ID.
  \item \code{HD} The product H × D (mm^2).
  \item \code{volume} The true volume of sterile water in the bladder (ml).
}
}
\keyword{datasets}
