% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeSeries.R
\name{cal_time}
\alias{cal_time}
\title{Calendar time helper}
\usage{
cal_time(
  xwin = c("dtd", "wtd", "mtd", "qtd", "ytd", "ttm", "3 yr", "5 yr", "10 yr", "20 yr",
    "30 yr"),
  asof = Sys.Date()
)
}
\arguments{
\item{xwin}{string to represent calendar window}

\item{asof}{ending date of window}
}
\value{
beginning date of window
}
\description{
Calendar time helper
}
