% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_put_calendar_spread_config.R
\name{calculate_calendar_score}
\alias{calculate_calendar_score}
\title{Calculate Opportunity Score}
\usage{
calculate_calendar_score(
  iv_rank = NA,
  iv_ratio = NA,
  net_theta = NA,
  net_vega = NA,
  debit_profit_ratio = NA,
  open_interest = NA,
  is_range_bound = TRUE,
  dividend_safe = TRUE,
  config = PUT_CALENDAR_SPREAD_CONFIG
)
}
\arguments{
\item{iv_rank}{IV Rank (0-1)}

\item{iv_ratio}{Front IV / Back IV ratio}

\item{net_theta}{Net theta (should be positive)}

\item{net_vega}{Net vega (should be positive)}

\item{debit_profit_ratio}{Debit to estimated max profit ratio}

\item{open_interest}{Minimum OI of both legs}

\item{is_range_bound}{Boolean indicating if stock is range-bound}

\item{dividend_safe}{Boolean indicating if ex-div timing is safe}

\item{config}{Configuration object}
}
\value{
Numeric score from 0-100
}
\description{
Calculates a composite score (0-100) for a calendar spread opportunity
based on the research framework scoring system.
}
