% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_money_market_rotation.R
\name{calculate_returns_by_period_type}
\alias{calculate_returns_by_period_type}
\title{Calculate returns by period type for table display}
\usage{
calculate_returns_by_period_type(
  rotations,
  dividends,
  capital_flows,
  period_type = "yearly"
)
}
\arguments{
\item{rotations}{Rotation history}

\item{dividends}{Dividend history}

\item{capital_flows}{Capital flow history}

\item{period_type}{One of "monthly", "quarterly", "yearly"}
}
\value{
Data frame with period returns
}
\description{
Calculate returns by period type for table display
}
