% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_money_market_rotation.R
\name{calculate_returns_by_year}
\alias{calculate_returns_by_year}
\title{Calculate returns by year}
\usage{
calculate_returns_by_year(rotations, dividends, capital_flows)
}
\arguments{
\item{rotations}{Rotation history}

\item{dividends}{Dividend history}

\item{capital_flows}{Capital flow history}
}
\value{
Data frame with yearly performance metrics
}
\description{
Breaks down performance metrics by calendar year.
}
