% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_money_market_rotation.R
\name{calculate_twr}
\alias{calculate_twr}
\title{Calculate time-weighted return}
\usage{
calculate_twr(rotations, dividends, capital_flows)
}
\arguments{
\item{rotations}{Rotation history}

\item{dividends}{Dividend history with rotation_id}

\item{capital_flows}{Inferred capital flows}
}
\value{
List with TWR percentage and sub-period details
}
\description{
Uses Modified Dietz method for TWR calculation with capital flows.
}
