% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_cash_secured_puts_config.R
\docType{data}
\name{CASH_SECURED_PUTS_CONFIG}
\alias{CASH_SECURED_PUTS_CONFIG}
\title{Cash-Secured Puts Configuration Constants}
\format{

}
\usage{
CASH_SECURED_PUTS_CONFIG
}
\description{
Configuration object containing all strategy-specific parameters for
cash-secured put analysis.
}
\keyword{datasets}
