% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeSeries.R
\name{combine_xts}
\alias{combine_xts}
\title{Combine multiple xts objects into one xts object}
\usage{
combine_xts(
  ...,
  period = NULL,
  dtype = c("return", "price"),
  use_busday = TRUE,
  comm_start = TRUE,
  comm_end = TRUE,
  na_rpl = TRUE,
  input_list = FALSE
)
}
\arguments{
\item{period}{optional string to specify a desired common frequency}

\item{dtype}{string to specify if we are combining 'return' or 'price' 
time-series}

\item{use_busday}{boolean to strip time-series to business day only
(NYSE calendar)}

\item{comm_start}{boolean to truncate the combined time-series to the 
latest inception of the individual xts objects}

\item{comm_end}{boolean to truncate the combined time-series to the earliest
ending date of the individual xts objects}

\item{na_rpl}{boolean to replace missing values to \code{0} for returns and
the lagged non-missing value for prices}

\item{input_list}{boolean to specify if a list of xts objects are being 
past, see vignette}
}
\value{
the combined xts object
}
\description{
Combine multiple xts objects into one xts object
}
