% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_risk_presets.R
\name{compare_risk_presets}
\alias{compare_risk_presets}
\title{Display preset comparison table}
\usage{
compare_risk_presets()
}
\value{
Data frame comparing presets
}
\description{
Helper for users to understand differences between presets
}
