% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ui_components.R
\name{create_groups_summary_alert}
\alias{create_groups_summary_alert}
\title{Create a summary alert for broken groups}
\usage{
create_groups_summary_alert(broken_count = 0, incomplete_count = 0)
}
\arguments{
\item{broken_count}{Number of broken groups}

\item{incomplete_count}{Number of incomplete groups}
}
\value{
HTML div element with alert styling
}
\description{
Creates a dismissible alert showing count of groups needing attention
}
