\docType{data}
\name{crystal}
\alias{crystal}
\title{Crystal weight data}
\format{A data frame with 14 rows and 2 variables}
\description{
  The data give the growing time and final weight of
  crystals.
}
\details{
  \itemize{ \item time time taken to grow (hours) \item
  weight final weight of the crystal (grams) }
}
\references{
  Graybill, F. A., and Iyer, H. K. Regression analysis:
  Concepts and Applications. Belmont, Calif: Duxbury Press,
  1994.
}
\keyword{datasets}

