% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeSeries.R
\name{.return_xts}
\alias{.return_xts}
\title{Return xts with index of dates}
\usage{
.return_xts(x)
}
\arguments{
\item{x}{xts object}
}
\description{
Return xts with index of dates
}
\note{
When using functions such as apply on xts objects an xts object with
a numeric index is returned, however the character dates are still in the
objects attributes. This function reconstructs the xts with a date index.
}
