% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownloadData.R
\name{download_fred}
\alias{download_fred}
\title{Download time-series from FRED database}
\usage{
download_fred(ticker, try_mode = TRUE)
}
\arguments{
\item{ticker}{string of time-series to download}

\item{try_mode}{boolean to use try-error handling}
}
\value{
xts containing dates and economic data
}
\description{
Download time-series from FRED database
}
\details{
see https://fred.stlouisfed.org for more info
}
\examples{
indust_prod <- download_fred('INDPRO')
}
