% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownloadData.R
\name{download_french}
\alias{download_french}
\title{Download data with daily frequency from Ken French datalibrary}
\usage{
download_french(data_url, skip)
}
\arguments{
\item{data_url}{subset of url specific to data, see details}

\item{skip}{how many header rows in csv file to skip, see details}
}
\value{
xts with time-series
}
\description{
Download data with daily frequency from Ken French datalibrary
}
\details{
the base url for all downloads is
'http://mba.tuck.dartmouth.edu/pages/faculty/ken.french/ftp/', each
csv download has a unique url to add after the base, e.g., the \code{data_url}
for momentum is 'F-F_Momentum_Factor_daily_CSV.zip'. The amount of rows to 
skip is different for each dataset. You can check by downloading the csv in
the datalibray https://mba.tuck.dartmouth.edu/pages/faculty/ken.french/data_library.html
and seeing how many header rows above the data there are. As of 11/1/2020 some
common skips are: daily 3 factor = 4, daily momentum = 12, daily 10 sector = 9.
}
