% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownloadData.R
\name{download_tiingo}
\alias{download_tiingo}
\title{Download data from Tiingo}
\usage{
download_tiingo(ticker, t_api, try_mode = TRUE)
}
\arguments{
\item{ticker}{string to represent security ticker}

\item{t_api}{string to represent your Tiingo API, see details}

\item{try_mode}{boolean to use try / catch expression}
}
\value{
xts of historical returns
}
\description{
Download data from Tiingo
}
\details{
Tiingo provies end of day prices for stocks, ETFs, and mutual funds.
You can set up a free account to get an API. www.tiingo.com
}
