% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_etf_covered_calls_config.R
\name{etf-screener-config}
\alias{etf-screener-config}
\title{ETF Screener Covered Calls Strategy Configuration}
\description{
Configuration constants for the ETF screener covered call strategy
}
