% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_collar_etf_universe.R
\name{etf-subsets}
\alias{etf-subsets}
\alias{get_index_etfs}
\alias{get_sector_etfs}
\alias{get_bond_etfs}
\alias{get_commodity_etfs}
\title{Get specific ETF subsets}
\usage{
get_index_etfs()

get_sector_etfs()

get_bond_etfs()

get_commodity_etfs()
}
\value{
Character vector of ETF tickers
}
\description{
Helper functions to get specific types of ETFs
}
\section{Functions}{
\itemize{
\item \code{get_index_etfs()}: Get major index ETFs only

\item \code{get_sector_etfs()}: Get sector ETFs only

\item \code{get_bond_etfs()}: Get bond ETFs only

\item \code{get_commodity_etfs()}: Get commodity ETFs only

}}
