% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeSeries.R
\name{excess_ret}
\alias{excess_ret}
\title{Subtract risk-free from returns}
\usage{
excess_ret(x, rf, period = NULL)
}
\arguments{
\item{x}{xts object of returns}

\item{rf}{xts object with the risk-free return time-series}

\item{period}{optional period to change the frequency, default value of NULL
will not change the frequency and leave the time-series as is}
}
\value{
xts object containing \code{x} less \code{rf}
}
\description{
Subtract risk-free from returns
}
