% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_explainability.R
\name{explainability}
\alias{explainability}
\title{Explainability UI and Logic}
\description{
Provides functions to generate UI components and server logic for
explaining complex risk metrics in a user-friendly way.
}
