% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeSeries.R
\name{fill_na_price}
\alias{fill_na_price}
\title{Fill missing price values with previous non-missing value}
\usage{
fill_na_price(x)
}
\arguments{
\item{x}{xts with prices}
}
\value{
\code{x} with missing values replaced by the previous non-missing value
}
\description{
Fill missing price values with previous non-missing value
}
\note{
see https://stackoverflow.com/questions/7735647/replacing-nas-with-latest-non-na-value
}
