% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeSeries.R
\name{fill_na_ret}
\alias{fill_na_ret}
\title{Fill missing values for a return time-series}
\usage{
fill_na_ret(x, rpl = 0)
}
\arguments{
\item{x}{xts with returns}

\item{rpl}{value to replace missing values}
}
\value{
\code{x} with NAs replaced by \code{rpl}
}
\description{
Fill missing values for a return time-series
}
\details{
Missing values at the beginning of the time-series are presevered.
Only the missing values after a non-missing value are replaced. E.g., the
series NA, NA, 1, 5, NA, 7 would be replaced with NA, NA, 1, 5, 0, 7.
}
