% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_formatting.R
\name{format_date}
\alias{format_date}
\title{Format date consistently}
\usage{
format_date(date, format = "\%Y-\%m-\%d")
}
\arguments{
\item{date}{Date object or character date}

\item{format}{Date format string (default "%Y-%m-%d")}
}
\value{
Character string formatted date
}
\description{
Format date consistently
}
\examples{
\dontrun{
  format_date(Sys.Date())                      # "2025-09-30"
  format_date(Sys.Date(), "\%b \%d, \%Y")         # "Sep 30, 2025"
}
}
