% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_formatting.R
\name{format_number}
\alias{format_number}
\title{Format large number with abbreviations}
\usage{
format_number(value, digits = 1)
}
\arguments{
\item{value}{Numeric value to format}

\item{digits}{Number of decimal places (default 1)}
}
\value{
Character string with abbreviated number
}
\description{
Converts large numbers to K (thousands), M (millions), B (billions)
}
\examples{
\dontrun{
  format_number(1234)           # "1.2K"
  format_number(1234567)        # "1.2M"
  format_number(1234567890)     # "1.2B"
}
}
