% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_formatting.R
\name{format_time_ago}
\alias{format_time_ago}
\title{Format timestamp as time ago}
\usage{
format_time_ago(timestamp)
}
\arguments{
\item{timestamp}{POSIXct timestamp}
}
\value{
Character string with relative time
}
\description{
Converts a timestamp to human-readable relative time (e.g., "5 mins ago")
}
\examples{
\dontrun{
  format_time_ago(Sys.time() - 60)           # "1 min ago"
  format_time_ago(Sys.time() - 3600)         # "1 hour ago"
  format_time_ago(Sys.time() - 86400)        # "1 day ago"
}
}
