% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_formatting.R
\name{format_with_commas}
\alias{format_with_commas}
\title{Format number with thousands separator}
\usage{
format_with_commas(value, digits = 0)
}
\arguments{
\item{value}{Numeric value to format}

\item{digits}{Number of decimal places (default 0)}
}
\value{
Character string with comma separators
}
\description{
Simple formatting with commas for thousands
}
\examples{
\dontrun{
  format_with_commas(1234567)        # "1,234,567"
  format_with_commas(1234.56, 2)     # "1,234.56"
}
}
