% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeSeries.R
\name{freq_to_scaler}
\alias{freq_to_scaler}
\title{Convert string frequency to numeric scale}
\usage{
freq_to_scaler(period)
}
\arguments{
\item{period}{string to represent frequency}
}
\value{
numeric scale
}
\description{
Convert string frequency to numeric scale
}
