% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_money_market_rotation.R
\name{get_available_periods}
\alias{get_available_periods}
\title{Get available periods for filter}
\usage{
get_available_periods(rotations)
}
\arguments{
\item{rotations}{Rotation history}
}
\value{
List with period options
}
\description{
Get available periods for filter
}
