% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_options_cache.R
\name{get_cache_stats}
\alias{get_cache_stats}
\title{Get cache statistics}
\usage{
get_cache_stats()
}
\value{
List with cache statistics
}
\description{
Returns information about current cache state.
}
