% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_money_market_rotation.R
\name{get_current_mm_position}
\alias{get_current_mm_position}
\title{Get current position summary}
\usage{
get_current_mm_position(rotations)
}
\arguments{
\item{rotations}{Rotation history}
}
\value{
List with current ticker, shares, cost basis, or NULL if no position
}
\description{
Get current position summary
}
