% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_custom_ticker_lists.R
\name{get_drip_tickers}
\alias{get_drip_tickers}
\title{Get ticker symbols from DRIP Investing by dividend category}
\usage{
get_drip_tickers(category)
}
\arguments{
\item{category}{One of "King", "Champion", "Contender", "Challenger"}
}
\value{
Character vector of ticker symbols
}
\description{
Get ticker symbols from DRIP Investing by dividend category
}
