% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_money_market_rotation.R
\name{get_first_week_date}
\alias{get_first_week_date}
\title{Get first WEEK transaction date}
\usage{
get_first_week_date(conn)
}
\arguments{
\item{conn}{DuckDB connection}
}
\value{
Date or NA if no WEEK transactions
}
\description{
Get first WEEK transaction date
}
