% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_collar_etf_universe.R
\name{get_index_etfs}
\alias{get_index_etfs}
\title{Get major index ETFs only}
\usage{
get_index_etfs()
}
\value{
Character vector of ETF tickers
}
\description{
Get major index ETFs only
}
\seealso{
Other etf-subset-helpers: 
\code{\link{get_bond_etfs}()},
\code{\link{get_commodity_etfs}()},
\code{\link{get_sector_etfs}()}
}
\concept{etf-subset-helpers}
