% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_money_market_rotation.R
\name{get_mm_rotation_summary}
\alias{get_mm_rotation_summary}
\title{Get rotation summary statistics}
\usage{
get_mm_rotation_summary(rotations, dividends)
}
\arguments{
\item{rotations}{Rotation history}

\item{dividends}{Dividend history}
}
\value{
List with summary stats
}
\description{
Get rotation summary statistics
}
