% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_background_refresh.R
\name{get_refresh_status}
\alias{get_refresh_status}
\title{Get the current refresh status}
\usage{
get_refresh_status()
}
\value{
List with status information
}
\description{
Returns information about the last successful and failed refreshes
for both activities and positions data.
}
