% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_collar_etf_universe.R
\name{get_sector_etfs}
\alias{get_sector_etfs}
\title{Get sector ETFs only}
\usage{
get_sector_etfs()
}
\value{
Character vector of ETF tickers
}
\description{
Get sector ETFs only
}
\seealso{
Other etf-subset-helpers: 
\code{\link{get_bond_etfs}()},
\code{\link{get_commodity_etfs}()},
\code{\link{get_index_etfs}()}
}
\concept{etf-subset-helpers}
