% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_risk_config.R
\name{get_stress_scenario}
\alias{get_stress_scenario}
\title{Get stress scenario definition}
\usage{
get_stress_scenario(scenario_name)
}
\arguments{
\item{scenario_name}{Scenario key (e.g., "financial_crisis_2008")}
}
\value{
Scenario definition list
}
\description{
Get stress scenario definition
}
