% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_activity_linking.R
\name{get_unlinked_by_symbol}
\alias{get_unlinked_by_symbol}
\title{Get unlinked activities by symbol}
\usage{
get_unlinked_by_symbol()
}
\value{
Tibble with symbol and count of unlinked activities
}
\description{
Groups unlinked activities by symbol to make manual review easier.
}
