% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getData}
\alias{getData}
\alias{getData.lm}
\title{Extract Data from an Object}
\usage{
\method{getData}{lm}(object)
}
\arguments{
\item{object}{An object inheriting from class \code{lm}, representing a
fitted linear model.}
}
\value{
if a \code{data} argument is present in the calling sequence that
        produced \code{object}, the corresponding data frame (with
        \code{na.action} and \code{subset} applied to it, if also present in
        the call that produced \code{object}) is returned; else, \code{NULL}
         is returned.
}
\description{
If present in the calling sequence used to produce \code{object}, the data
frame used to fit the model is obtained.
}

