% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getNewData}
\alias{getNewData}
\title{Default Predictor Values}
\usage{
getNewData(x)
}
\description{
Sets the default predictor values for \code{invest} when \code{newdata} is
missing.
}
\keyword{internal}

