% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getVarInfo}
\alias{getVarInfo}
\title{Extract Variables Information from a Fitted Model}
\usage{
getVarInfo(object, data)
}
\arguments{
\item{object}{A fitted model object.}
}
\value{
A list containing the following components:
\describe{
  \item{\code{x.names}}{The names of the predictor variables.}
  \item{\code{x.names}}{The names of the response variables.}
  \item{\code{x}}{The values of the predictor variables.}
  \item{\code{y}}{The values of the response variables.}
  \item{\code{x.dim}}{The number of predictor variables used.}
  \item{\code{y.dim}}{The number of response variables used.}
}
}
\description{
Extract the variables used to fit a particular model.
}

