% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_activity_linking.R
\name{ignore_activity}
\alias{ignore_activity}
\title{Mark activity to be ignored for grouping}
\usage{
ignore_activity(activity_id)
}
\arguments{
\item{activity_id}{Activity identifier}
}
\value{
Logical TRUE if successful
}
\description{
Sets ignore_for_grouping flag for activities that should not be linked to groups.
This is a convenience wrapper around the batch function.
}
