% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_money_market_rotation.R
\name{infer_capital_flows}
\alias{infer_capital_flows}
\title{Infer capital flows from rotation history}
\usage{
infer_capital_flows(rotations, dividends)
}
\arguments{
\item{rotations}{Data frame from build_rotation_history}

\item{dividends}{Data frame of dividends with rotation_id}
}
\value{
Data frame with inferred additions/withdrawals
}
\description{
Compares expected cash (sell proceeds + dividends) to next buy amount
to determine capital additions or withdrawals.
}
