% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_activity_linking.R
\name{link_activities_to_group}
\alias{link_activities_to_group}
\title{Link multiple activities to a group}
\usage{
link_activities_to_group(activity_ids, group_id)
}
\arguments{
\item{activity_ids}{Character vector of activity IDs}

\item{group_id}{Group identifier}
}
\value{
Logical TRUE if all successful
}
\description{
Batch operation to link multiple activities to the same group.
}
