% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_activity_linking.R
\name{link_activity_to_group}
\alias{link_activity_to_group}
\title{Link activity to group}
\usage{
link_activity_to_group(activity_id, group_id)
}
\arguments{
\item{activity_id}{Activity identifier}

\item{group_id}{Group identifier}
}
\value{
Logical TRUE if successful
}
\description{
Updates the group_id for a specific activity.
}
