% Generated by roxygen2 (4.0.2): do not edit by hand
\name{makeData}
\alias{makeData}
\title{Make new data frame}
\usage{
makeData(x, label)
}
\description{
Create a new data frame from a specified x value that has the same structure
as the data frame used to create \code{object}. (For internal use only.)
}
\keyword{internal}

