% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{makeX}
\alias{makeX}
\title{Construct design matrix for fixed effects}
\usage{
makeX(object, newdata)
}
\description{
Create a fixed effects design matrix from \code{newdata} based on a fitted
model. (For internal use only.)
}
\keyword{internal}

