% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_token_settings.R
\name{mod_token_settings_server}
\alias{mod_token_settings_server}
\title{Token Settings Server}
\usage{
mod_token_settings_server(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\description{
Token Settings Server
}
