% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_token_settings.R
\name{mod_token_settings_ui}
\alias{mod_token_settings_ui}
\title{Token Settings UI}
\usage{
mod_token_settings_ui(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\value{
Shiny UI elements
}
\description{
Token Settings UI
}
