% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_monthly_dividend_config.R
\name{monthly-dividend-config}
\alias{monthly-dividend-config}
\title{Monthly Dividend Stocks Strategy Configuration}
\description{
Configuration constants for the monthly dividend stocks covered call strategy
}
