% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dividend_capture_monthly_config.R
\docType{data}
\name{MONTHLY_ETFS}
\alias{MONTHLY_ETFS}
\title{Monthly Dividend ETFs List}
\format{
A tibble with 76 rows and 3 columns:
\describe{
  \item{ticker}{ETF ticker symbol}
  \item{description}{ETF description/name}
  \item{schedule_type}{Dividend schedule pattern}
}
}
\usage{
MONTHLY_ETFS
}
\description{
List of 76 monthly dividend ETFs for analysis
}
\keyword{datasets}
