% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{nasturtium}
\alias{nasturtium}
\title{Bioassay on Nasturtium}
\format{A data frame with 42 rows and 2 variables}
\source{
Racine-Poon, A. (1988) A Bayesian Approach to Nonlinear Calibration
Problems, \emph{Journal of the American Statistical Association}, \bold{83},
650--656.
}
\description{
The data give the actual concentrations of an agrochemical present in soil
samples versus the weight of the plant after three weeks of growth.
}
\details{
\itemize{
  \item \code{conc} True concentration of agrochemical (g/ha).
  \item \code{weight} Weight of plant (mg) after 3 weeks' growth.
}
}
\references{
Huet, S., Bouvier, A., Poursat, M-A., and Jolivet, E.  (2004)
\emph{Statistical Tools for Nonlinear Regression: A Practical Guide with
S-PLUS and R Examples}. Springer.
}
\keyword{datasets}

