% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page_monthly_dividend.R
\name{page_monthly_dividend}
\alias{page_monthly_dividend}
\title{Monthly Dividend Stocks Page}
\usage{
page_monthly_dividend()
}
\value{
A brochure page object
}
\description{
Monthly Dividend Stocks Page
}
