% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_money_market_rotation.R
\name{parse_period_selection}
\alias{parse_period_selection}
\title{Parse period selection to date range}
\usage{
parse_period_selection(period_value, rotations)
}
\arguments{
\item{period_value}{Value from period selector}

\item{rotations}{Rotation history (for "all" option)}
}
\value{
List with start_date and end_date
}
\description{
Parse period selection to date range
}
