% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_questrade_healthcheck.R
\name{ping_questrade}
\alias{ping_questrade}
\title{Ping Questrade API to rotate refresh token}
\usage{
ping_questrade()
}
\value{
Logical TRUE if successful, FALSE otherwise
}
\description{
Silent health check that authenticates with Questrade and automatically
saves the new refresh token. Keeps token alive as long as app runs at
least once every 7 days.
}
