% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/invest.R
\name{plot.bootCal}
\alias{plot.bootCal}
\title{Plots of the Output of a Bootstrap Calibration Simulation}
\usage{
\method{plot}{bootCal}(x, ...)
}
\arguments{
\item{x}{An object that inherits from class \code{"bootCal"}.}

\item{...}{Additional optional arguments. At present, no optional arguments
are used.}
}
\description{
This takes a bootstrap calibration object and produces plots for the
bootstrap replicates of the inverse estimate.
}

