% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_portfolio_groups_config.R
\name{portfolio-groups-config}
\alias{portfolio-groups-config}
\title{Portfolio Groups Configuration Utilities}
\description{
Helper functions for accessing position groups configuration
from golem-config.yml
}
