% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/predFit.R
\name{predFit}
\alias{predFit}
\alias{predFit.lm}
\alias{predFit.lme}
\alias{predFit.nls}
\title{Predictions from a Fitted Model}
\usage{
predFit(object, ...)

\method{predFit}{lm}(object, newdata, se.fit = TRUE, interval = c("none",
  "confidence", "prediction"), level = 0.95, adjust = c("none",
  "Bonferroni", "Scheffe"), k, ...)

\method{predFit}{nls}(object, newdata, se.fit = TRUE, interval = c("none",
  "confidence", "prediction"), level = 0.95, adjust = c("none",
  "Bonferroni", "Scheffe"), k, ...)

\method{predFit}{lme}(object, newdata, se.fit = TRUE, ...)
}
\description{
Generic prediction method for various types of fitted models. (For internal
use only.)
}
\keyword{internal}

