% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pspline.R
\name{predict.pspline}
\alias{predict.pspline}
\title{Predicting from Penalized Regression Spline Fits}
\usage{
\method{predict}{pspline}(object, newdata, se.fit = FALSE,
  interval = c("none", "confidence", "prediction"), level = 0.95)
}
\description{
Generic prediction method for various types of fitted models. (For internal 
use only.)
}
\keyword{internal}

