% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/predict2.R
\name{predict2}
\alias{predict2}
\title{Predict method for (Single-Regressor) Linear, Nonlinear, and (Linear) Mixed
Model Fits}
\usage{
predict2(object, ...)
}
\description{
Generic prediction method for various types of fitted models. (For internal
use only.)
}
\keyword{internal}

