% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeSeries.R
\name{price_to_ret}
\alias{price_to_ret}
\title{Convert prices to returns (percent change)}
\usage{
price_to_ret(x)
}
\arguments{
\item{x}{xts of prices}
}
\value{
returns calculated from \code{x}
}
\description{
Convert prices to returns (percent change)
}
