% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pspline.R
\name{pspline}
\alias{print.pspline}
\alias{pspline}
\alias{pspline.data.frame}
\alias{pspline.default}
\alias{pspline.matrix}
\title{Fitting Penalized Regression Splines}
\usage{
pspline(x, ...)

\method{pspline}{default}(x, y, degree = 1, knots, num.knots,
  method = c("REML", "ML"), ...)

\method{pspline}{data.frame}(x, ...)

\method{pspline}{matrix}(x, ...)
}
\arguments{
\item{x}{A numeric vector giving the values of the explanatory variable.}

\item{...}{Additional optional arguments to be passed on to \code{nlme::lme}.}

\item{y}{A numeric vector giving the values of the response variable.}

\item{degree}{Integer specifying the degree of the piecewise polynomial, 
default is \code{1} for piecewise linear splines.}

\item{knots}{A numeric vector specifying the internal breakpoints that define 
the spline.}

\item{num.knots}{Integer specifying the number of knots.}

\item{method}{Method used for comuting the variance components. Default is
\code{"REML"}.}
}
\value{
An object of class \code{pspline} (essentially a list) containing the 
  following components:
  \describe{
    \item{\code{x}}{The vector of predictor values.}
    \item{\code{y}}{The vector of response values.}
    \item{\code{xname}}{The name of the predictor variable.}
    \item{\code{yname}}{The name of the respons variable.}
    \item{\code{fit}}{The vecotr of fitted values.}
    \item{\code{knots}}{Numeric vector containing the knot locations.}
    \item{\code{degree}}{The degree of the penalized regression spline.}
    \item{\code{beta.hat}}{Vector of estimated fixed effects.}
    \item{\code{u.hat}}{Vector of estimated random effects.}
    \item{\code{spar}}{The smoothing parameter.}
    \item{\code{var.rat}}{The ratio of variance components.}
    \item{\code{matrices}}{A list of key matrices.}
    \item{\code{var.components}}{Vector of variance components.}
    \item{\code{df.res}}{Residual degrees of freedom.}
    \item{\code{lmm}}{The original \code{"lme"} object.}
    \item{\code{call}}{The (matched) function call.}
  }
}
\description{
Fits a penalized regression spline to the supplied data using a linear mixed- 
effects model representation.
}
\examples{
# Proof of whisky stored in a charred oak barrel against time in years (see 
# Schoeneman, et.al. 1971)
whiskey <- data.frame(
  age = c(0, 0.5, 1, 2, 3, 4, 5, 6, 7, 8),
  proof = c(104.6, 104.1, 104.4, 105.0, 106.0, 
            106.8, 107.7, 108.7, 110.6, 112.1)
)
plot(whiskey)
whiskey_ps <- with(whiskey, pspline(age, proof, degree = 1))
plotFit(whiskey_ps)
}
\references{
Ruppert, D., Wand, M. & Carroll, R. (2003). Semiparametric regression. 
Cambridge New York: Cambridge University Press.
}

