% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_put_calendar_spread_config.R
\docType{data}
\name{PUT_CALENDAR_SPREAD_CONFIG}
\alias{PUT_CALENDAR_SPREAD_CONFIG}
\title{Put Calendar Spread Configuration Constants}
\format{

}
\usage{
PUT_CALENDAR_SPREAD_CONFIG
}
\description{
Configuration object containing all strategy-specific parameters for
put calendar spread analysis on dividend aristocrats.
}
\keyword{datasets}
