% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_questrade_quotes.R
\name{questrade-quotes}
\alias{questrade-quotes}
\title{Questrade Quote Fetching Functions}
\description{
Functions for fetching stock quotes from Questrade API with Yahoo Finance
compatible output format.
}
