% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_background_refresh.R
\name{refresh_questrade_activities}
\alias{refresh_questrade_activities}
\title{Refresh Questrade Activities}
\usage{
refresh_questrade_activities()
}
\value{
A promise that resolves with the result of fetch_all_activities(),
  which is a tibble of fetched activities.
}
\description{
Fetches the latest account activities from Questrade API and saves them
to the database. This function is designed to be called from a background
process or a timer.
}
