% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_risk_presets.R
\name{risk-presets}
\alias{risk-presets}
\title{Risk Analysis Preset Configurations}
\description{
Defines preset configurations for different user types, following KISS principle.
90% of users will use these presets without customization.
}
