% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_options_cache.R
\name{set_cached_options}
\alias{set_cached_options}
\title{Store options chain in cache}
\usage{
set_cached_options(ticker, options_data)
}
\arguments{
\item{ticker}{Stock ticker symbol}

\item{options_data}{List containing options chain data}
}
\description{
Stores options chain data in session cache with timestamp.
}
