% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/utilities.R
\name{Sigma}
\alias{Sigma}
\title{Extract residual standard error}
\usage{
Sigma(object, ...)
}
\description{
Extract residual standard error from a fitted model. (For internal use
only.)
}
\keyword{internal}

