% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_risk_config.R
\docType{data}
\name{STRESS_SCENARIOS}
\alias{STRESS_SCENARIOS}
\title{Pre-built Stress Test Scenarios}
\format{
An object of class \code{list} of length 5.
}
\usage{
STRESS_SCENARIOS
}
\description{
Based on historical crises and academic research
}
\keyword{datasets}
