% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_risk_config.R
\docType{data}
\name{TICKER_SECTOR_MAP}
\alias{TICKER_SECTOR_MAP}
\title{Map tickers to sectors (for stress testing)}
\format{
An object of class \code{list} of length 59.
}
\usage{
TICKER_SECTOR_MAP
}
\description{
Common sector classifications. Expand as needed.
}
\keyword{datasets}
