% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeSeries.R
\name{trunc_xts}
\alias{trunc_xts}
\title{Truncate xts object}
\usage{
trunc_xts(x, date_start = NULL, date_end = NULL)
}
\arguments{
\item{x}{xts object}

\item{date_start}{first date to truncate}

\item{date_end}{last date to truncate}
}
\description{
Truncate xts object
}
