% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_activity_linking.R
\name{unlink_activity}
\alias{unlink_activity}
\title{Unlink activity from group}
\usage{
unlink_activity(activity_id)
}
\arguments{
\item{activity_id}{Activity identifier}
}
\value{
Logical TRUE if successful
}
\description{
Sets group_id to NULL for an activity (useful for corrections).
}
