% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{varY}
\alias{varY}
\title{Evaluate response variance}
\usage{
varY(object, newdata)
}
\description{
Evaluate response variance at a given value of the predictor variable. (For 
internal use only.)
}
\keyword{internal}

