% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataViz.R
\name{viz_drawdown}
\alias{viz_drawdown}
\title{Plot Drawdowns}
\usage{
viz_drawdown(x, period = NULL)
}
\arguments{
\item{x}{xts object}

\item{period}{optional string to change the frequency of the xts}
}
\description{
Plot Drawdowns
}
