% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataViz.R
\name{viz_perf}
\alias{viz_perf}
\title{Plot Performance Summary}
\usage{
viz_perf(x, period = NULL, plot_col = NULL)
}
\arguments{
\item{x}{xts object}

\item{period}{optional string to change the frequency of the xts}
}
\description{
Plot Performance Summary
}
