% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dividend_capture_weekly_config.R
\docType{data}
\name{WEEKLY_ETFS}
\alias{WEEKLY_ETFS}
\title{Weekly Dividend ETFs List}
\format{
A tibble with 32 rows and 2 columns:
\describe{
  \item{ticker}{ETF ticker symbol}
  \item{description}{ETF description/name}
}
}
\usage{
WEEKLY_ETFS
}
\description{
List of 32 weekly dividend ETFs for analysis
}
\keyword{datasets}
