% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_explainability.R
\name{with_info_icon}
\alias{with_info_icon}
\title{Add an info icon with a tooltip}
\usage{
with_info_icon(ui_element, tooltip_text, placement = "top")
}
\arguments{
\item{ui_element}{The Shiny UI element to add the icon to.}

\item{tooltip_text}{The text to display in the tooltip.}

\item{placement}{The placement of the tooltip (e.g., "top", "right").}
}
\value{
A Shiny UI element.
}
\description{
A helper function to wrap a UI element with an info icon that displays
a tooltip on hover.
}
