\docType{data}
\name{arsenic}
\alias{arsenic}
\title{Concentrations of arsenic in water samples}
\format{A data frame with 32 rows and 2 variables}
\description{
The data give the actual and measures concentration of
arsenic present in water samples.
}
\details{
\itemize{ \item actual True amount of arsenic present \item
measured Measured amount of arsenic present }
}
\references{
Graybill, F. A., and Iyer, H. K. Regression analysis:
Concepts and Applications. Belmont, Calif: Duxbury Press,
1994.
}
\keyword{datasets}

