\docType{data}
\name{arsenic}
\alias{arsenic}
\title{Concentrations of arsenic in water samples.}
\format{A data frame with 32 rows and 2 variables}
\description{
  The data give the actual and measures concentration of
  arsenic present in water samples. \itemize{ \item actual
  True amount of arsenic present \item measured Measured
  amount of arsenic present }
}
\references{
  \url{http://www.stat.colostate.edu/~hari/regression_book/index.html}
}
\keyword{datasets}

