% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{arsenic}
\alias{arsenic}
\title{Concentrations of arsenic in water samples}
\format{A data frame with 32 rows and 2 variables}
\source{
Graybill, F. A., and Iyer, H. K. (1994)
\emph{Regression analysis: Concepts and Applications}. Duxbury Press.
}
\description{
The data give the actual and measured concentrations of arsenic present in
water samples.
}
\details{
\itemize{
  \item \code{actual} True amount of arsenic present.
  \item \code{measured} Measured amount of arsenic present.
}
}
\keyword{datasets}
