% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crystal}
\alias{crystal}
\title{Crystal weight data}
\format{A data frame with 14 rows and 2 variables}
\source{
Graybill, F. A., and Iyer, H. K. (1994)
\emph{Regression analysis: Concepts and Applications}. Duxbury Press.
}
\description{
The data give the growing time and final weight of crystals.
}
\details{
\itemize{
  \item \code{time} Time taken to grow (hours).
  \item \code{weight} Final weight of the crystal (grams).
}
}
\keyword{datasets}

