\docType{data}
\name{crystal}
\alias{crystal}
\title{Crystal weight data.}
\format{A data frame with 14 rows and 2 variables}
\description{
  The data give the growing time and final weight of
  crystals. \itemize{ \item time time taken to grow (hours)
  \item weight final weight of the crystal (grams) }
}
\references{
  \url{http://www.stat.colostate.edu/~hari/regression_book/index.html}
}
\keyword{datasets}

