% Generated by roxygen2 (4.0.2): do not edit by hand
\name{invest}
\alias{invest}
\alias{invest.lm}
\alias{invest.lme}
\alias{invest.nls}
\title{Calibration for Linear and Nonlinear Regression Models.}
\usage{
invest(object, ...)

\method{invest}{lm}(object, y0, interval = c("inversion", "Wald", "none"),
  level = 0.95, mean.response = FALSE, data, pboot = FALSE, nsim = 1,
  seed = NULL, lower, upper, tol = .Machine$double.eps^0.25,
  maxiter = 1000, adjust = c("none", "Bonferroni"), k, ...)

\method{invest}{nls}(object, y0, interval = c("inversion", "Wald", "none"),
  level = 0.95, mean.response = FALSE, data, boot = FALSE,
  type = c("parametric", "nonparametric"), nsim = 1, seed = NULL,
  progress = FALSE, lower, upper, tol = .Machine$double.eps^0.25,
  maxiter = 1000, adjust = c("none", "Bonferroni"), k, ...)

\method{invest}{lme}(object, y0, interval = c("inversion", "Wald", "none"),
  level = 0.95, mean.response = FALSE, data, lower, upper, q1, q2,
  tol = .Machine$double.eps^0.25, maxiter = 1000, ...)
}
\arguments{
\item{object}{An object that inherits from class \code{"lm"}, \code{"nls"},
or \code{"lme"}.}

\item{...}{Additional optional arguments. At present, no optional arguments
are used.}

\item{y0}{The value of the observed response(s) or specified value of the
mean response.}

\item{interval}{The type of interval required.}

\item{level}{A numeric scalar between 0 and 1 giving the confidence level for
the interval to be calculated.}

\item{mean.response}{Logical indicating whether confidence intervals should
              correspond to an individual response (\code{FALSE}) or a
              mean response
(\code{TRUE}).}

\item{data}{An optional data frame. This is required if \code{object$data}
is \code{NULL}.}

\item{pboot}{Logical indicating whether to perform a parametric bootstrap.}

\item{nsim}{Number of bootstrap simulations, positive integer; the bootstrap
B (or R).}

\item{seed}{Optional argument to set.seed.}

\item{lower}{The lower endpoint of the interval to be searched.}

\item{upper}{The upper endpoint of the interval to be searched.}

\item{tol}{The desired accuracy passed on to \code{uniroot}. Recommend a
minimum of 1e-10.}

\item{maxiter}{The maximum number of iterations passed on to \code{uniroot}.}

\item{adjust}{A logical value indicating if an adjustment should be made to
the critical value used in calculating the confidence interval.
This is useful for when the calibration curve is to be used
multiple, say k, times.}

\item{k}{The number times the calibration curve is to be used for computing a
confidence interval. Only needed when \code{adjust = "Bonferroni"}.}

\item{q1}{Optional lower cutoff to be used in forming confidence intervals.
Only used when object inherits from class \code{"lme"}. Defaults to
\code{qnorm((1+level)/2)}.}

\item{q2}{Optional upper cutoff to be used in forming confidence intervals.
Only used when object inherits from class \code{"lme"}. Defaults to
\code{qnorm((1-level)/2)}.}
}
\value{
If \code{boot = FALSE}, then an object of class \code{"calibrate"}
        containing the following components:
  \itemize{
    \item \code{estimate} The estimate of x0.
    \item \code{lwr} The lower confidence limit for x0.
    \item \code{upr} The upper confidence limit for x0.
    \item \code{se} An estimate of the standard error (Wald interval only).
    \item \code{interval} The method used for calculating \code{lower} and
          \code{upper} (only used by \code{print} method).
  }
        Otherwise, an object of class \code{"bootCal"} containing the
        following components:
  \itemize{
    \item \code{original} The estimate of x0.
    \item \code{bootreps} The lower confidence limit for x0.
    \item \code{nsim} The number of bootstrap replicates.
    \item \code{level} The desired confidence level.
  }
}
\description{
The function \code{invest} computes the inverse estimate and a condfidence
interval for the unknown predictor value that corresponds to an observed
value of the response (or vector thereof) or specified value of the mean
response. See the references listed below for more details.
}
\examples{
data(Puromycin, package = "datasets")
Puromycin2 <- Puromycin[Puromycin$state == "treated", ]
Puro2.nls <- nls(rate ~ (theta1 * conc) / (theta2 + conc),
                 data = Puromycin2, start = list(theta1 = 200, theta2 = 1))
plotFit(Puro2.nls, interval = "both")
invest(Puro2.nls, y0 = 100, interval = "inversion")
invest(Puro2.nls, y0 = 100, interval = "inversion", mean.response = TRUE)
}
\references{
Greenwell, B. M., and Schubert Kabban, C. M. (2014). investr: An R Package
for Inverse Estimation. \emph{The R Journal}, \bold{6}(1), 90--100.
URL http://journal.r-project.org/archive/2014-1/greenwell-kabban.pdf.

Graybill, F. A., and Iyer, H. K. (1994).
\emph{Regression analysis: Concepts and Applications}. Duxbury Press.

Huet, S., Bouvier, A., Poursat, M-A., and Jolivet, E.  (2004)
\emph{Statistical Tools for Nonlinear Regression: A Practical Guide with S-PLUS and R Examples}. Springer.

Seber, G. A. F., and Wild, C. J. (1989)
\emph{Nonlinear regression}. Wiley.

Oman, Samuel D. (1998).
Calibration with Random Slopes.
\emph{Biometrics} \bold{85}(2): 439--449.
doi:10.1093/biomet/85.2.439.
}

