\name{investr-package}
\alias{investr-package}
\alias{investr}
\docType{package}
\title{
Inverse Estimation in R
}
\description{
This package provides functions for inverse estimation with linear and nonlinear 
regression models in R.
}
\details{
\tabular{ll}{
Package: \tab investr\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-08-07\cr
License: \tab GPL (>= 2)\cr
}

}
\author{
Brandon M. Greenwell

Maintainer: Brandon M. Greenwell <brandon.greenwell@afit.edu>

}
\references{
Bates, D. M., and Watts, D. G. Nonlinear Regression Analysis and its 
Applications. New York: Wiley, 2007. 

Graybill, F. A., and Iyer, H. K. Regression analysis: Concepts and Applications. Belmont, Calif: Duxbury Press, 1994. 

Huet, S., Bouvier, A., Poursat, M-A., and Jolivet, E. Statistical Tools for 
Nonlinear Regression: A Practical Guide with S-PLUS and R Examples. New York: 
Springer, 2004. 

Seber, G. A. F., and Wild, C. J.. Nonlinear regression. New York: Wiley, 1989.

F. Baty and M. L. Delignette-Muller (2012), nlstools: tools for nonlinear 
regression.
}



