% Generated by roxygen2 (4.0.1): do not edit by hand
\name{makeData}
\alias{makeData}
\title{Make new data frame}
\usage{
makeData(object, x)
}
\description{
Create a new data frame from a specified x value that has the same structure
as the data frame used to create \code{object}.
}
\keyword{internal}

