% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{makeX}
\alias{makeX}
\alias{makeZ}
\title{Design Matrix for Fixed and Random Effects}
\usage{
makeX(object, newdata)

makeZ(object, newdata)
}
\description{
Construct the fixed or random effects design matrix from a fitted model using
\code{newdata}. (For internal use only.)
}
\keyword{internal}

