% Generated by roxygen2 (4.0.2): do not edit by hand
\name{makeZ}
\alias{makeZ}
\title{Construct design matrix for random effects}
\usage{
makeZ(object, newdata)
}
\description{
Create a random effects design matrix from \code{newdata} based on a fitted
model. (For internal use only.)
}
\keyword{internal}

