% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.bootCal}
\alias{plot.bootCal}
\title{Plots of the Output of a Bootstrap Calibration Simulation}
\usage{
\method{plot}{bootCal}(object)
}
\arguments{
\item{object}{An object that inherits from class \code{"bootCal"}.}
}
\description{
This takes a bootstrap calibration object and produces plots for the
bootstrap replicates of the inverse estimate.
}

