% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_questrade_healthcheck.R
\name{questrade-healthcheck}
\alias{questrade-healthcheck}
\title{Questrade API Health Check}
\description{
Minimal health check to keep Questrade refresh token alive.
Automatically rotates tokens on every ping.
}
