% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_risk_config.R
\docType{data}
\name{RISK_CONFIG}
\alias{RISK_CONFIG}
\title{Risk Analysis Configuration List}
\format{
An object of class \code{list} of length 25.
}
\usage{
RISK_CONFIG
}
\description{
Risk Analysis Configuration List
}
\keyword{datasets}
